requirements.txt
Flask==2.2.2
requests==2.24.0
builtwith==1.3.4
beautifulsoup4==4.12.2
pyOpenSSL==23.2.0
